[DUBBO-3137]: move more constants out of ConfigConstants (#4092)
