+++ dubbo-common/src/main/java/org/apache/dubbo/common/constants/RegistryConstants.java
@@ -133,4 +133,13 @@ public interface RegistryConstants {
      * key for router type, for e.g., "script"/"file",  corresponding to ScriptRouterFactory.NAME, FileRouterFactory.NAME
      */
     String ROUTER_KEY = "router";
+
+    /**
+     * The key name for export URL in register center
+     */
+    String EXPORT_KEY = "export";
+    /**
+     * The key name for reference URL in register center
+     */
+    String REFER_KEY = "refer";
 }
