+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/Constants.java
@@ -113,4 +113,5 @@ public interface Constants {
 
     String REGISTRIES_SUFFIX = "dubbo.registries.";
 
+    String ZOOKEEPER_PROTOCOL = "zookeeper";
 }
