+++ dubbo-configcenter/dubbo-configcenter-api/src/main/java/org/apache/dubbo/configcenter/Constants.java
@@ -14,35 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.dubbo.configcenter;
 
-package org.apache.dubbo.common.constants;
-
-/**
- * ConfigConstants
- */
-public interface ConfigConstants {
-    String CLUSTER_KEY = "cluster";
-
-    String USERNAME_KEY = "username";
-
-    String PASSWORD_KEY = "password";
-
-    String HOST_KEY = "host";
-
-    String PORT_KEY = "port";
-
-    String DUBBO_IP_TO_BIND = "DUBBO_IP_TO_BIND";
-
-    @Deprecated
-    String SHUTDOWN_WAIT_SECONDS_KEY = "dubbo.service.shutdown.wait.seconds";
-
-    String SHUTDOWN_WAIT_KEY = "dubbo.service.shutdown.wait";
-
-    String DUBBO_PROTOCOL = "dubbo";
-
-    String QOS_ENABLE = "qos.enable";
-
-    String QOS_PORT = "qos.port";
-
-    String ACCEPT_FOREIGN_IP = "qos.accept.foreign.ip";
+public interface Constants {
+    String CONFIG_CLUSTER_KEY = "config.cluster";
+    String CONFIG_NAMESPACE_KEY = "config.namespace";
+    String CONFIG_GROUP_KEY = "config.group";
+    String CONFIG_CHECK_KEY = "config.check";
 }
