+++ dubbo-registry/dubbo-registry-api/src/main/java/org/apache/dubbo/registry/Constants.java
@@ -15,34 +15,8 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.common.constants;
+package org.apache.dubbo.registry;
 
-/**
- * ConfigConstants
- */
-public interface ConfigConstants {
-    String CLUSTER_KEY = "cluster";
-
-    String USERNAME_KEY = "username";
-
-    String PASSWORD_KEY = "password";
-
-    String HOST_KEY = "host";
-
-    String PORT_KEY = "port";
-
-    String DUBBO_IP_TO_BIND = "DUBBO_IP_TO_BIND";
-
-    @Deprecated
-    String SHUTDOWN_WAIT_SECONDS_KEY = "dubbo.service.shutdown.wait.seconds";
-
-    String SHUTDOWN_WAIT_KEY = "dubbo.service.shutdown.wait";
-
-    String DUBBO_PROTOCOL = "dubbo";
-
-    String QOS_ENABLE = "qos.enable";
-
-    String QOS_PORT = "qos.port";
-
-    String ACCEPT_FOREIGN_IP = "qos.accept.foreign.ip";
+public interface Constants {
+    String REGISTER_IP_KEY = "register.ip";
 }
