+++ dubbo-remoting/dubbo-remoting-api/src/main/java/org/apache/dubbo/remoting/telnet/support/TelnetHandlerAdapter.java
@@ -26,7 +26,7 @@ import org.apache.dubbo.remoting.telnet.TelnetHandler;
 import org.apache.dubbo.remoting.transport.ChannelHandlerAdapter;
 
 import static org.apache.dubbo.common.constants.CommonConstants.COMMA_SPLIT_PATTERN;
-import static org.apache.dubbo.common.constants.ConfigConstants.TELNET;
+import static org.apache.dubbo.remoting.Constants.TELNET;
 
 public class TelnetHandlerAdapter extends ChannelHandlerAdapter implements TelnetHandler {
